SpicyTracker 2.0 remove all params no regex
Switched away from using regex, to instead have it try to turn it into a URL object. Then remove all the search parameters, I don't believe Spotify uses query parameters, except for tracking, on any of their copy actions, so this should work, please make an issue if it doesn't. It also catches and return the original text if it fails.